Move pre-allocated storage classes to separate headers
