+++ test/header.cpp
@@ -1,7 +1,8 @@
 #include <doctest/doctest.h>
 #include <openssl/crypto.h>
 #include <openssl/err.h>
-#include <sframe/sframe.h>
+
+#include <header.h>
 
 #include "common.h"
