+++ test/vectors.cpp
@@ -3,6 +3,7 @@
 #include <nlohmann/json.hpp>
 #include <sframe/sframe.h>
 
+#include <header.h>
 #include <crypto.h>
 
 #include "common.h"
